refactor: 도메인 페키지 이동
