+++ src/main/java/com/soolsul/soolsulserver/post/business/PostCommandService.java
@@ -4,8 +4,8 @@ import com.soolsul.soolsulserver.auth.User;
 import com.soolsul.soolsulserver.post.domain.Post;
 import com.soolsul.soolsulserver.post.domain.PostRepository;
 import com.soolsul.soolsulserver.post.presentation.dto.PostCreateRequest;
-import com.soolsul.soolsulserver.restaurant.Restaurant;
-import com.soolsul.soolsulserver.restaurant.RestaurantRepository;
+import com.soolsul.soolsulserver.restaurant.domain.Restaurant;
+import com.soolsul.soolsulserver.restaurant.domain.RestaurantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
