+++ src/main/java/com/soolsul/soolsulserver/restaurant/domain/RestaurantMoodTag.java
@@ -1,4 +1,4 @@
-package com.soolsul.soolsulserver.restaurant;
+package com.soolsul.soolsulserver.restaurant.domain;
 
 import com.soolsul.soolsulserver.common.domain.BaseEntity;
 import lombok.AccessLevel;
@@ -9,7 +9,7 @@ import javax.persistence.Entity;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RestaurantAlcoholTag extends BaseEntity {
+public class RestaurantMoodTag extends BaseEntity {
 
     @Column(nullable = false)
     private String restaurantId;
@@ -17,8 +17,14 @@ public class RestaurantAlcoholTag extends BaseEntity {
     @Column(nullable = false)
     private String name;
 
-    public RestaurantAlcoholTag(String restaurantId, String name) {
+    private boolean isCuration;
+
+    private Integer count;
+
+    public RestaurantMoodTag(String restaurantId, String name, boolean isCuration, Integer count) {
         this.restaurantId = restaurantId;
         this.name = name;
+        this.isCuration = isCuration;
+        this.count = count;
     }
 }
